+++ lib/Settings/Admin.php
@@ -91,7 +91,9 @@ class Admin implements ISettings {
 				'encryption_enabled_for_groupfolders' => $this->config->getAppValue('groupfolders', 'enable_encryption', '') === 'true'
 			],
 			'oidc_provider' => $this->openProjectAPIService->getRegisteredOidcProviders(),
-			'user_oidc_enabled' => $this->openProjectAPIService->isUserOIDCAppInstalledAndEnabled()
+			'user_oidc_enabled' => $this->openProjectAPIService->isUserOIDCAppInstalledAndEnabled(),
+			'user_oidc_supported' => $this->openProjectAPIService->isUserOIDCAppSupported(),
+			'user_oidc_minimum_version' => OpenProjectAPIService::MIN_SUPPORTED_USER_OIDC_APP_VERSION,
 		];
 
 		$this->initialStateService->provideInitialState('admin-settings-config', $adminConfig);