+++ src/constants/messages.js
@@ -16,4 +16,6 @@ export const messages = {
 
 export const messagesFmt = {
 	appNotInstalled: (app) => t(APP_ID, 'The "{app}" app is not installed', { app }),
+	appNotSupported: (app) => t(APP_ID, 'The "{app}" app is not supported', { app }),
+	minimumVersionRequired: (minimumAppVersion) => t(APP_ID, 'Requires app version "{minimumAppVersion}" or later', { minimumAppVersion }),
 }