+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Show errors when user_oidc app is not available [#753](https://github.com/nextcloud/integration_openproject/pull/753)
 - Show proper error message in the dashboard based on auth method [#770](https://github.com/nextcloud/integration_openproject/pull/770)
 - Drop application's support for Nextcloud 27 [#779](https://github.com/nextcloud/integration_openproject/pull/779)
+- Show error when the user_oidc app not supported [#768](https://github.com/nextcloud/integration_openproject/pull/768)
 
 ### Fixed
 - choose correct base URL for OCS requests [#780](https://github.com/nextcloud/integration_openproject/pull/780)